Remove unclear wording: info descriptor values
All of the sections describing info descriptors had a sentence like
"Every <SYCL object> must must produce a valid value for each info
parameter", but it was not clear what this means.  In fact, some info
descriptors are defined to throw exceptions in some circumstances
(e.g. `info::device::parent_device`), so this is an example of when
an info descriptor does not return a valid value.

Since these sentences do not seem to add any clarity, I removed them.