Clarify kernel_device_specific::global_work_size
This info descriptor can only be used for built-in kernels or for a
"custom" device.  I think the intent is that each back-end
specification will clarify what it really means, so note that in its
description.

This clarification was provoked by a question from our QA team because
they were confused about the expected functionality.