Clarification to is_compatible()
Clarify that `is_compatible()` returns false for a kernel that is
decorated with `[[sycl::requires(has(attr))]]` if the device does not
support `attr`, even if the kernel does not actually use a feature that
corresponds to `attr`.

This is consistent with the behavior when submitting such a kernel to
a device, which throws `kernel_not_supported` even if the kernel does
not actually use a feature that corresponds to `attr`.