Add read-write host unsampled image accessor
A reviewer noted to me that SYCL 1.2.1 supported a read-write image
accessor for `target::host_image`.  SYCL 2020 inadvertently dropped
this functionality because `host_unsampled_image_accessor` does not
support read-write access mode.  Add that functionality back, so we
are again on parity with SYCL 1.2.1.

Adding a read-write mode also has a happy side-effect.  This is now
a reasonable default mode, so change the synopsis to show read-write
as the default (unless `dataT` is `const`).  Adding a default for
`accessMode` also allows us to add a default for `dimensions`, so
make that default to `1`, which matches the default for
`unsampled_image` and the default for the other accessors.