Local accessor should not have is_placeholder
The deprecated form of `accessor` with `target::local` did not have a
member function named `is_placeholder` in SYCL 1.2.1 because there is
no such thing as a local placeholder accessor.  The SYCL 2020 spec
inadvertently added `is_placeholder` as part of the deprecated API,
which was not our intent.  Remove it.

This addresses one of the issues reported in #126.