Standard attribute position for non-lambda kernels
We used to define the position of C++ attributes attached to kernel
functions such that they appertained to the "function-type" of the
kernel function.  This is necessary for lambda functions because the
C++17 standard does not define an attribute position that appertains to
the lambda function itself.  (This is expected to be fixed in C++23.)

However, non-lambda functions have a well-defined position for
attributes that appertain to the function itself.  It's not clear why
the spec did not use this position for these functions.

This commit changes the position of these attributes to the standard
place.

I also made some clarifications about the general restrictions of
these attributes, which were suggested by our front-end experts.  I
think these do not change the intent, but just tighten the wording.