Remove old attribute form
Remove the deprecated C++ attribute forms that are inherited from
SYCL 1.2.1.  Many details of these attributes were not well defined,
and we thought it was better to just remove them now rather than
spending time hashing out the details of a feature that we plan to
remove soon anyways.  We expect that existing implementations will
continue to support these attributes with their existing semantics,
but we encourage applications to migrate to the new (better defined)
forms as soon as possible.

Closes internal issue 553.