Invert order of src/dest parameters for USM copy
We noticed that the USM `copy()` functions had a parameter order of
`(dest, src)`, which is opposite of the `copy()` functions for
accessors.  This seemed very confusing, so we decided to change the
order to match.  Since these USM functions are new in SYCL 2020 and
because no major implementation has added them yet, we think it is
still safe to make this change.

This closes internal issue #557