Rename "sycl::requires()" to "sycl::device_has()"
Make two changes to the `[[sycl::requires()]]` attribute based on some
implementation experience:

* Change the name from `requires` to `device_has` in order to avoid
  confusion with the C++20 "requires" keyword.

* Simplify the parameter syntax by removing the `has()` clause.  The
  parameter list is now just a list of aspects.

The second bullet point deserves some explanation.  Our original intent
was to allow the `[[sycl::requires()]]` attribute to express both
conjunctions and disjunctions of aspects in some future version of the
SYCL spec.  For example, we thought we might want to allow syntax like
this in the future:

```
[[sycl::requires(has(aspect::fp64, aspect::atomic64) || has(aspect::fp16))]]
```

We still think this might be desirable in the future, but we think
disjunctions like this will be very uncommon.  We don't want to incur
the complexity of the `has()` clause even in the common case, so we
omit it from the current syntax (which only supports conjunctions of
aspects).

If we do decide to support disjunction in the future, we can create a
new attribute that supports the expanded syntax.  For example:

```
[[sycl::device_has_expression(has(aspect::fp64, aspect::atomic64) || has(aspect::fp16))]]
```

(We can bikeshed the name `device_has_expression` later if we decide to
add this.)