Clarify the rules for legal kernel parameters
Update the section 4.12.4 "Rules for parameter passing to kernels" to
make the following clarifications:

* Clarify which variables are treated as kernel parameters.
* Clarify that an accessor (or other SYCL parameter type) can be passed
  as a member variable of some other class type.
* Clarify that an accessor (or other SYCL parameter type) can be passed
  as a base class of some other class type.

We used to have a list of SYCL types that could be passed as kernel
parameters in section 4.4 "Class availability".  Move this list into
section 4.12.4 instead to keep all the information about kernel
parameter types in one place.

Closes internal issue #225