Clarify reduction behavior for mismatched identity
The original text said that supplying an identity value different
to the identity associated with the reduction operator was undefined
behavior, which was too strong; the reduction is still expected to
create, initialize and eventually combine all of the required reducer
objects.

The key point is that the result of the reduction is not guaranteed
in this case, which is expressed here by weaker "unspecified" phrasing.
A non-normative note is attached to clarify this behavior for users.