Fix definition of fp_config::correctly_rounded_divide_sqrt
The definition that was used before was taken from the OpenCL 1.2
specification, see section 4.2, table 4.3:

> CL_FP_CORRECTLY_ROUNDED_DIVIDE_SQRT – divide and sqrt are correctly
> rounded as defined by the IEEE754 specification.

However this definition is misleading because what is really advertised
by this property is that the correctly rounded divide and sqrt are
supported, not that they are correctly rounded by default on the
given device.

In the OpenCL 1.2 specification this is made clear in section 5.6.4.2
which defines the build program flag
`-cl-fp32-correctly-rounded-divide-sqrt`.

Where it states the following:

> If this build option is not specified, the minimum numerical accuracy of
> single precision floating-point divide and sqrt are as defined in section 7.4
> of the OpenCL specification.

And:

> This build option can only be specified if the
> CL_FP_CORRECTLY_ROUNDED_DIVIDE_SQRT is set in CL_DEVICE_SINGLE_FP_CONFIG (as
> defined in table 4.3) for devices that the program is being build.
> clBuildProgram or clCompileProgram will fail to compile the program for a
> device if the -cl-fp32-correctly-rounded-divide-sqrt option is specified and
> CL_FP_CORRECTLY_ROUNDED_DIVIDE_SQRT is not set for the device.

But since SYCL doesn't specify build options, this entire section doesn't
appear in the SYCL specification and we're left with just the misleading
property definition that makes it sound like the correctly rounded divide and
sqrt will be used by default.

And so this patch alters the wording to only say that the correctly rounded
variants are supported, and not suggest that they are always used.