Deprecate device property correctly_rounded_divide_sqrt
This property was taken from the matching OpenCL property, however, the
OpenCL property is there to advertise support for
`-cl-fp32-correctly-rounded-divide-sqrt`. And since the SYCL
specification doesn't define anything like that this device property
isn't nearly as useful.

This feature can be provided by vendors through environment variable sor
compiler flags, and for devices using the OpenCL backend, it can be
accessed with the OpenCL backend interop.