Fix prototype of OCL make_event
Fix the prototype of the OpenCL backend interop function `make_event()`.
The type of the first parameter must be
`const backend_input_t<backend::opencl, event> &`, but this type alias
is defined as `std::vector<cl_event>` in the table "List of native
types per SYCL object in the OpenCL backend".  We decided that the
table is correct, so change the prototype to match.

Closes internal issue 588.