Fix description of atomic_ref operators
Prior to this commit, the atomic_ref operators were incorrectly
described as being equivalent to calling a fetch_* member function
(e.g. += was described as being equivalent to calling fetch_add).

In ISO C++, fetch_* functions return the value before the atomic
operation is performed, and operators return the value after the atomic
operation is performed.