Clarify empty accessor
Even before this change, we specified that the accessor default
constructor creates an empty accessor.  Previously, we said

> The only iterator that can be obtained is nullptr.

However, this makes no sense because iterators aren't (necessarily)
pointers.  I considered changing this to say that `begin() == `end()`,
however, this is already implied because we say that `empty()` is
`true`.  I thought it was more important to say that `get_pointer()`
and `get_multi_ptr()` return unspecified values in this case.  I also
clarified the descriptions of those two functions for the empty
accessor case.