Remove confusing paragraph about events
Remove this paragraph about obtaining an event for a command group.
It's not clear what this paragraph is trying to communicate, and some
of the information is wrong.  For example, you cannot get an event that
represents "start of the command group function object" or for the
"kernel starting".  You can only get a single event that represents the
command group as a whole.  You can get profiling information that tells
when these things happen, but you cannot get an event.

Section 4.6.6 "Event class" already describes events and the profiling
information they provide.  Section 3.7.1.2 "SYCL command groups and
execution order" already describes the execution model of command
groups and their relationship to the buffers they use.  Therefore,
there is no new information being presented in this paragraph.

Together with #242 close internal issue 424.