Clarify event profiling information
Clarify what the three event profiling timestamps mean.  Also clarify
which timestamps share the same "timebase", which provides some
guarantee about which timestamps can be compared in order to calculate
a time duration.

The event profiling information is related to the event status.  I
noticed that there was no description of the possible statuses of an
event, so I added a table for this.

Closes internal issue 424