Remove const from operator[] in reducer
Prior to this commit, the subscript operator of reducer was marked
const due to a copy-paste error from the definition in accessor.

Marking this operator const prevents the reducers associated with
sycl::span reductions from storing data directly, forcing them to
be implemented as an accessor-like view. This is inconsistent with
the requirements of other (scalar) reductions; the definitions of
combine() and other shorthand operators (e.g. ++) aren't marked const.