Clarify behavior of queue destructor
The spec already said that the `queue` destructor does not block in
section 3.9.8.1 "Synchronization in the SYCL application":

> Note that the destructors of other SYCL objects (sycl::queue,
> sycl::context,...) do not block.

However, we never really said what happens to outstanding commands
that have been submitted to the queue.  This PR clarifies that these
commands just run normally, with the same behavior as though the
queue had not been destroyed.

This seems like a sensible behavior, and it also enables a use case
similar to `std::thread::detach`.