Stable order of platform and device objects
Clarify that the order of the `platform` and `device` objects is the
same each time `get_platforms()` or `get_devices()` is called.  This
allows an application to store the index of a platform or device and
use that index later to identify the same platform or device.

Also clarify that the `get_devices()` function returns only the root
devices, not any sub-devices that have been created.

Closes internal issue 603.