Fix for PDF build problem vs. asciidoctor-pdf
Per
https://github.com/KhronosGroup/SYCL-Docs/pull/268#issuecomment-1198790889

A recent update to the asciidoctor-spec Docker image moved it from
asciidoctor-pdf 1.6.1 to 2.2.0, and deprecated the `pdf-stylesdir`
attribute, causing the PDF build to fail. This branch updates the
deprecated attribute and adds a note to the repository README.

There are still some warnings about the custom PDF theme, which should
be updated to match the 2.2.0 theme, but this is the minimal change to
get CI functioning again. I'll work on the theme shortly as well.