Small clarification to buffer sync rules
The section on buffer synchronization rules define the scenarios when
the buffer destructor must block vs. the scenarios where it does not
block.  The old wording made it seem like a buffer created from no host
pointer is only non-blocking if it uses the default memory allocator.
I assume this was not our intent, and such a buffer should have a
non-blocking destructor regardless of the memory allocator.

The main change in this PR is to remove the phrase "and using the
default buffer allocator".  However, I also reworded a bit to clarify
which type of buffers we are talking about.