Buffer from contain always blocks in destructor
Clarify the behavior of the buffer destructor for a buffer that is
created from a container.  The previous wording said that the
destructor blocks if the container is non-const, but it didn't say
what happens for a const container.  The new wording clarifies that it
blocks in this case also.

This is consistent with the behavior of buffers that are constructed
from a host pointer.  In that case, the destructor always blocks, even
when host pointer is const.