Clarify buffer creation with nullptr
Clarify that the overloads of the buffer constructor that take a
`hostData` pointer do not allow `nullptr` unless the buffer size is
zero.

This was already pretty clear for the buffer constructors that take a
raw pointer.  However, there was a statement indicating that the
overloads which take a `std::shared_ptr` could be called with an
"empty" pointer (i.e. null).  This seemed inconsistent.  Why should we
allow null for the form with `std::shared_ptr` but not for the form
with raw pointers?  Applications that want the SYCL runtime to allocate
space for the buffer should just call the overload that takes no
`hostData` pointer.