Move description of shared_ptr buffer destructor
The buffer destructor has special rules about blocking when the buffer
is constructed from a `shared_ptr`.  Previously, these rules were
described in section 4.7.4.3 "Shared SYCL ownership of the host memory".
However, the behavior of the buffer destructor in other cases is
described in section 4.7.2.3 "Buffer synchronization rules".  This PR
moves the description to 4.7.2.3 in order to keep the descriptions of
the buffer destructor all in one place.

After doing this, I realized that section 4.7.4.3 did not contain any
new information that wasn't already described in 4.7.2.3.  Therefore, I
removed that section.

I also tightened the specification a bit for the case when a buffer is
constructed from a `shared_ptr`:

* If the underlying type of the `shared_ptr` is `const`, then I assume
  we want the buffer to be read-only, just as in the raw pointer case.

* I also assume that the contents of the host memory referenced by the
  `shared_ptr` is undefined during the lifetime of the buffer, just as
  in the raw pointer case.