Clarify "group" meaning in algorithm descriptions
The specification sometimes uses "group" to mean "work-group" (as in
sycl::group) and at other times to mean "group of work-items" (a concept
describing both sycl::group and sycl::sub_group).

This commit clarifies several usages of "group" in the descriptions of
group algorithms, by explicitly referring to the specific instance of
a group object (g).