Clarify the "reducer" member types and constants
This was motivated by some confusion that arose while reviewing the
CTS test plan for reduction variables.  It wasn't clear to some
reviewers what the dimensionality of the reduction variable was in all
cases.  This commit adds a table which clarifies the value of
`dimensions` as well as the type aliases that are defined for the
reducer class.