Allow "empty" shared_ptr for buffer construction
During review we decided that it should be allowed to construct a
buffer from an empty `shared_ptr`.  Restore the wording I removed in
the previous commit and clarify that empty `shared_ptr` is allowed in
all the relevant constructor overloads.

Note that null raw pointers are still not allowed, consistent with the
previous wording.

I also clarified the constructors taking `shared_ptr` to note that the
application can safely release ownership of the pointer after the
constructor returns.  This is consistent with the existing wording and
example in section 4.7.4.3 "Shared SYCL ownership of the host memory".

In order to avoid confusion, I changed wording like:

> The ownership of this memory is given to the constructed SYCL buffer

To:

> The buffer assumes exclusive access to this memory

This seems better because the word "ownership" has a different meaning
for `shared_ptr`.

Similar changes were also made to the image APIs.