Clarify sycl::atomic_ref
The main change in this commit is to limit `sycl::atomic_ref` to be
used only in device code and to provide a guarantee that
`std::atomic_ref` will be atomic w.r.t. `sycl::atomic_ref` when the
device supports host-to-device atomics.  This follows our decision in
internal issue 637.

It also provides some other clarifications:

* Describes the effect of the `AddressSpace` template parameter.
* Clarifies that the type `T` must either be 32-bits or 64-bits.  If it
  is 64-bits, the device must have `aspect::atomic64`.

It also removes a suggestion that the compiler should issue a
diagnostic if device code calls `std::atomic` or `std::atomic_ref`.  It
seems odd to provide this suggestion only for these two functions since
the SYCL spec does not allow device code to call any standard C++
function.  If we want to provide a suggestion (or requirement) like
this, I think it should be a more comprehensive statement about calling
standard C++ functions in general from device code.

Closes internal issue 637.