Reword guarantee about host-to-device fence sync
Clarify that the fence operations are defined in terms of atomic
operations on some shared atomic object.  This is defined in the core
C++ specification, but it seems useful to mention it here also to help
a reader understand the paragraphs in this section.  We still have
pointers to the C++ specification for a full description of the fence
semantics.