Disallow "++" and "--" for "vec<bool>"
This issue was raised by the Intel contractors who are writing CTS
tests.

C++17 does not allow `++` or `--` (either prefix or postfix) for the
`bool` type, so presumably we do not want to allow these operators in
SYCL for the `vec<bool>` type either.  This change marks these
operators as disallowed.

C++ *does* allow other arithmetic and bitwise operators on `bool`, so
we do *not* disallow any of these operators on `vec<bool>`.  By
following normal C++ rules, `bool` is promoted to `int` for these
operators.  Therefore, an operation like
`vec &operator+=(vec&, const vec&)` would first promote each element
of the operands to `int`, perform the operation, and then use "boolean
conversion" to convert the result back to `bool`.

I did not make any equivalent change to `marray` because the change
required there is larger in scope.  See internal issue 559 for more
details.