Clarify "is_compatible"
This PR addresses an issue that was raised during the review of
KhronosGroup/SYCL-CTS#553.  It is also related to the clarification
made in #370.

When we clarified the semantics of `any_device_has` and
`all_devices_have`, we decided that these tell information about the
"compilation environment" used to compile the SYCL translation unit.
This implies that kernels in a TU might not be compatible with some
devices depending on the "compilation environment".  For example, a
vendor's compiler might support some compiler flag that allows kernels
to run on GPU devices but not on other devices.

While writing the CTS, we realized that applications need some way to
tell if a kernel was compiled in a way that restricts the devices on
which it can run.  It seems like the existing API `is_compatible` is
a good fit, so this PR extends the semantics to say that this API also
considers the "compilation environment" when determining compatibility.