Editorial change with "any_device_has_v"
We had feedback from a reviewer that our description of `any_device_has`
and `all_devices_have` is confusing because we sometimes refer to the
shortcuts with the "_v" suffix (`any_device_has_v` and
`all_devices_have_v`).  Avoid this confusion by consistently using the
trait names without the "_v".

The description of the example continues to use the "_v" names because
the example uses those shortcuts, so it seemed appropriate to explain
the code that is actually used in the example.