Merge pull request #343 from npmiller/relax-backend-macros
Relax requirement on backend traits being available.

The current wording mandates that backend_traits specializations must be provided for every backend in the backend enum class.

This seems a bit overly restrictive and means that implementations would have to change the backend enum class depending on which backends are being supported on a specific installation.

In addition backend traits are defined based on underlying vendor types, which may rely on specific vendor headers that are not always available.

So this patch allows implementations to use the backend macro to also cover the case where the backend is in the enum class but the backend traits and, or backend are still not available. This should work fine with existing applications since using the backend traits requires the enum value and so backend traits use should already be guarded by the backend macro.