Merge pull request #295 from gmlueck/gmlueck/buffer-nullptr
Clarify buffer creation with nullptr.

Clarify that buffer and image constructors that take shared_ptr
support an empty shared_ptr, which causes the buffer or image to have
uninitialized contents. There was previously a statement to this
effect in section 4.7.4.3 "Shared SYCL ownership of the host memory",
but the sections describing the constructors themselves didn't say
anything about this.

Null raw pointers are still not allowed, consistent with the previous
wording. However, this clarifies that null raw pointers are allowed when
the buffer size is also zero.

This also clarifies the constructors taking shared_ptr to note that the
application can safely release ownership of the pointer after the
constructor returns. This is consistent with the existing wording and
example in section 4.7.4.3 "Shared SYCL ownership of the host memory".

In order to avoid confusion, this changes wording like:

    The ownership of this memory is given to the constructed SYCL buffer

To:

    The buffer assumes exclusive access to this memory

This seems better because the word "ownership" has a different meaning
for shared_ptr.