Remove image deduction guides
SYCL 2020 images do not carry information about the data type as part of
the type. As a result of this, the mentioned deduction guides for the
tag constructors are not implementable, as they would not be able to
deduce the `DataT` template argument.

This commit removes the tag constructors from all image accessors with
them and changes the `get_access` and `get_host_access` member functions
as they will now need explicit template arguments.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>