Clarify "weak" / "strong", "narrow" / "wide"
Clarify what it means for a memory order to be "weak" or "strong".
Clarify what it means for a memory scope to be "narrow" or "wide".

There were a couple of places that used the term "broad" instead of
"wide" when referring to memory scopes.  Changes these to be
consistent.