Clarify default constructed accessor and "require"
The previous wording caused some confusion about whether a default
constructed `accessor` is a placeholder.  In general, the accessor
constructors that do not take a `handler` parameter construct
placeholder accessors, which could give the impression that the
default constructed accessor is a placeholder.  Clarify this by
specifically stating that the default constructed accessor is not a
placeholder and that it may be passed as a kernel parameter without
calling `handler::require`.

The default constructed `accessor` and the default constructed
`local_accessor` are similar because they can both be passed as
kernel parameters and neither allocates any memory for use in the
kernel.  Adjust the wording for these two constructors to be more
similar.

Remove wording from `handler::require` saying that it is an error to
call this function for a zero sized accessor.  We allow non-placeholder
accessor that are zero sized, so it makes no sense to disallow this for
placeholder accessors.

Remove wording from the default `accessor` constructor saying that it
is an error to pass such an accessor to `handler::require`.  We allow
other non-placeholder accessors to be passed to `handler::require`, so
it makes sense to allow this for the default constructed accessor too.

Clarify the wording for `accessor::is_placeholder`.  The previous
wording "if the accessor was constructed as a placeholder" caused
confusion.  What if an accessor was constructed as a placeholder and
then later assigned to a non-placeholder accessor (or swapped to a
non-placeholder accessor)?  Presumably, we want `is_placeholder` to
return `false` in such a case.  The new wording removes this confusing
phrasing.

Closes #408.