Merge pull request #427 from steffenlarsen/steffen/clarify_placeholder_capture_exception
Currently the specification mandates that an implementation throws a synchronous exception when a command is submitted that has captured an accessor that has not previously been through handler::require. However, it may be difficult for host_task and general commands in library-only implementations to guarantee. To alleviate these problems, this commit aims to relax these exception requirements make the synchronous exception optional and require select accessor member functions to trigger asynchronous exceptions if the corresponding command is executed.

Additionally, this commit adds notes to select member functions in the deprecated accessors specializations for target::local and target::constant specifying that they can only be used in commands. This corresponds to the same interfaces for the non-deprecated accessor classes.