Clarify the builtin functions (again)
After gaining some implementation experience, we decided to change the
clarification of the builtin functions.  Our previous attempt in #428
added many overloads, and we discovered that this increased the time it
takes to compile the <sycl/sycl.hpp> header.  After further reflection,
we decided that a different mix of overloads vs. templates was more
consistent with C++ and it also results in better compile times.

In general, we follow C++ whenever there is a precedent.  For example,
the math functions like `fmax` and `isinf` have separate overloads for
scalar inputs because C++ defines these same function in that way.
However, we use template parameters for the variants that take `marray`
or `vec` because there is no C++ precedent here.  Other functions
like `clamp` have a C++ precedent where scalar arguments are template
parameters, so SYCL follows suit here also.

Of course, there are many cases where there is no C++ precedent.  In
these cases, SYCL tries to be self consistent.  For example, the
relational functions like `isequal` have separate overloads for scalar
inputs to be consistent with `isinf` and the other math functions.

We also felt that the "generic type names" were adding confusion to
the spec.  Therefore, this commit remove all these generic pseudo-type
definitions and shows the exact synopsis for each builtin function.
This was hard to do with the old table format, so the tables defining
the builtin functions have been completely rewritten to match the
style used by the newer builtin functions (e.g. the group functions).
This style leaves more horizontal space for the synopsis, which makes
it easier to write the synopsis without awkward line breaks.  The
"Constraints" section of this style also makes it easier to describe
the constraints for the various template parameters.

Some other changes of note:

* The Common Functions were inconsistently specified for the `half`
  type.  We had defined some `half` overloads, but we missed others.
  This commit adds `half` support consistently to these functions.

* The introductory paragraphs have been modified to remove general
  statements about the types that are supported by the functions.
  These statements were sometime in conflict with the actual function
  definitions, and it seemed better to avoid redundancy.  Each function
  clearly specifies the allowed input types, so we do not need to say
  this again in the introduction.

* The "native precision" and "half precision" math functions have been
  split out into their own sub-sections.  It was easy to miss these
  functions before because they came after a long table of regular math
  functions.  It's now much easier to see that they exist.

* All uses of "latexmath" have been removed from the descriptions of
  the builtin functions and replaced with Asciidoc formatting.  This
  results in a more consistent style, and it also fixed several
  problems where the latexmath version was not rendered correctly in
  the final PDF and/or HTML documents.

Closes internal issue 278 (again).
Closes #321